requirements.txt
alembic==0.9.1
appdirs==1.4.3
bcolz==0.12.1
Bottleneck==1.2.0
cachetools==2.0.0
click==6.7
contextlib2==0.5.4
cookies==2.2.1
cyordereddict==1.0.0
Cython==0.25.2
DateTime==4.1.1
decorator==4.0.11
Django==1.10.5
django-bootstrap3==8.1.0
empyrical==0.2.2
et-xmlfile==1.0.1
intervaltree==2.1.0
jdcal==1.3
Logbook==1.0.0
lru-dict==1.1.6
Mako==1.0.6
MarkupSafe==1.0
multipledispatch==0.4.9
networkx==1.11
numexpr==2.6.2
numpy==1.12.0
openpyxl==2.4.5
packaging==16.8
pandas==0.18.1
pandas-datareader==0.3.0.post0
patsy==0.4.1
pbr==1.10.0
progressbar33==2.4

# can't just use pymssql until https://github.com/pymssql/pymssql/issues/432#issuecomment-238257092
git+https://github.com/pymssql/pymssql.git

pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
requests==2.13.0
requests-file==1.4.1
requests-ftp==0.3.1
scipy==0.19.0
six==1.10.0
sortedcontainers==1.5.7
SQLAlchemy==1.1.6
statsmodels==0.8.0
tables==3.3.0
testfixtures==4.13.4
toolz==0.8.2
xlrd==1.0.0
git+git://github.com/quantopian/zipline@5fd20e1
zope.interface==4.3.3
